+++ crates/theme_redox/src/lib.rs
@@ -1,0 +1,30 @@
+/*!
+
+This crate provides the redox theme resources of OrbTks redox theme dark and light.
+It provides fonts, icons and colors.
+
+ */
+
+use orbtk_theming::{config::ThemeConfig, prelude::*};
+
+pub mod prelude;
+
+use orbtk_theme_default::register_default_fonts;
+
+/// Resource file of redox theme
+pub const THEME_REDOX: &str = include_str!("../theme/theme_redox.ron");
+
+/// The redox dark theme colors resource file.
+pub const THEME_REDOX_COLORS_DARK: &str = include_str!("../theme/theme_redox_colors.ron");
+
+/// The font resources of the redox theme
+pub const THEME_REDOX_FONTS: &str = include_str!("../theme/theme_redox_fonts.ron");
+
+/// Returns the redox OrbTk theme.
+pub fn theme_redox() -> Theme {
+    register_default_fonts(Theme::from_config(
+        ThemeConfig::from(THEME_REDOX)
+            .extend(ThemeConfig::from(THEME_REDOX_COLORS_DARK))
+            .extend(ThemeConfig::from(THEME_REDOX_FONTS)),
+    ))
+}